stable channel
mostly unmodified copy from tailscale-k8s

borrowed from [kingdonb/tailscale-k8s](https://github.com/kingdonb/tailscale-k8s)

there is no "stable" semver release so for now, the stable channel
serves the prerelease (<1.0.0) builds.

When it is declared stable, the patch can be removed and the stable
channel will serve the stable release!

automation - support for stable, edge, prerelease

and latest-any via ImagePolicy

Other image policies could be added if more environments are needed. All
of our automation will write to the `main` branch, except the dev policy
which is meant to be used on forks. (Pretend this repo is hosted in an
org, and developers all have their own forks! The dev environments maybe
should generally be derived to one of those forks.)

Instead of alpha/beta/test/dev we have opted for an "edge" channel which
is similar to prerelease, but can have those breaking changes or partial
features implemented only as a pre-release even after the 1.0.0 stable.

* dev

The dev environment is for users in forks. It provides a way for dev
users to patch in their own git repository and is provided for example,
and for a place to target the `latest-any` image policy.

(It is not meant to be used in the base repo, as `latest-any` can be
very noisy and wouldn't make any sense to automate in the main branch.)

* example of use

and fix broken thing

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

* fix ingress mismatches

you should not be blocked from putting both automation and production on
the same cluster

set up river-auto to write to these

Signed-off-by: Kingdon Barrett <kingdon@weave.works>